font: tell one cached mask from another by the whole of its key
A cached mask is found by the caller's number for it together with the
transform it was rendered under, and both were held in C types whose
width is not the one they carry. The number is an integer object and sat
in a long, which is half the integer's width on some platforms: two
integers a build can tell apart arrived as one number there, and the
mask filed under either was served to the other. The transform was
quantized into the fixed-point field the face's own transforms use, by a
conversion given whatever the multiplication produced -- so a coefficient
past that field had no answer to be converted to, several such
transforms reached the same one, and each was served whichever of their
rasters was filed first. That was reachable at either width: a font
scaled past the field, drawing a glyph small enough to cache, shared one
entry across every scale above it.

The number is now carried in a field spanning the integer's, pinned to
it at compile time. The transform is quantized through a conversion that
is told the field's bounds, and a coefficient outside them has no
fixed-point form rather than one it shares: the mask is not keyed at
all, so it is built each time it is wanted and drawn from its own
description. The advances filed beside a mask, and the width a /Metrics
entry states, go through the same bounded conversion for the same
reason -- each is a number a program supplies, and neither has a value
to be reported as once it leaves the field.

Commit message generated by Claude (Anthropic).
Source code generated by Claude (Anthropic).

+++ src/lib/xpost_font.h
@@ -272,12 +272,12 @@ void xpost_font_cache_setparams(long bmax, long lower, long upper);
    the transform they were rendered under. It is not font machinery: a
    glyph is one kind of mask and its face and character code are one way
    of naming one. It lives here because fonts were its first caller. */
-int xpost_mask_cache_lookup(const void *k1, unsigned long k2,
+int xpost_mask_cache_lookup(const void *k1, unsigned long long k2,
                                  const long m[4], long size,
                                  unsigned char **bits, int *rows, int *width,
                                  int *pitch, int *left, int *top,
                                  long *advance_x, long *advance_y);
-int xpost_mask_cache_insert(const void *k1, unsigned long k2,
+int xpost_mask_cache_insert(const void *k1, unsigned long long k2,
                                  const long m[4], long size,
                                  const unsigned char *bits, int rows,
                                  int width, int pitch, int left, int top,
